Developer documentation

Tableau

MetrichHeap

Developer documentation

Tableau

SoleReasoners.AbstractTableau — Type

abstract type AbstractTableau end

Abstract type for all tableaux structures.

An analytic tableau is a tree structure computed for a logical formula having at each node a subformula of the original formula to be proved or refuted. It is used in many automated reasoning tasks, such as automated theorem proving and the satifiability problem.

SoleReasoners.Tableau — Type

mutable struct Tableau <: AbstractTableau
    const formula::Formula
    const father::Union{Tableau, Nothing}
    children::Vector{Tableau}
    expanded::Bool
    closed::Bool
end

A mutable structure representing a tableau as a tree structure, with each node containing a subformula of the original formula, the father and children in the tree structure, a flag saying if the node has already been expanded and a flag saying if the branch represented by the node has been closed. Each path from a leaf to the root respresents a branch.

SoleReasoners.ManyValuedTableau — Type

mutable struct ManyValuedTableau{T<:Truth} <: AbstractTableau
    const signedformula::SignedFormula{T}
    const father::Union{ManyValuedTableau{T}, Nothing}
    children::Vector{ManyValuedTableau{T}}
    expanded::Bool
    closed::Bool
end

A mutable structure representing a tableau as a tree structure, with each node containing a signed formula, the father and children in the tree structure, a flag saying if the node has already been expanded and a flag saying if the branch represented by the node has been closed. Each path from a leaf to the root respresents a branch.

SignedFormula

SoleReasoners.SignedFormula — Type

struct SignedFormula{T<:Truth}
    sign::Bool
    boundingimplication::Union{Tuple{T, Formula}, Tuple{Formula, T}, Tuple{T, T}}
end

All formulas appearing in the many-valued tableau will be bounding implications, i.e., a → A (or A → a), where a is a propositional constant and asserting a ≤ A (resp. A ≤ a).

MetrichHeap

SoleReasoners.MetricHeapNode — Type

struct MetricHeapNode{T<:AbstractTableau}
    metricvalue::Int
    tableau::T
end

The atomic element of a MetricHeap, it contains a tableau branch and a value for the metric associated with the MetricHeap it is contained in.

SoleReasoners.MetricHeap — Type

struct MetricHeap
    heap::BinaryHeap{MetricHeapNode}
    metric::Function
end

A MetricHeap is basically a heap parametrized over a metric, i.e., a function which extracts some information about a tableau branch, therefore containing in each node a tableau branch and the relative value for the metric, and which is ordered as a min heap over this metric value.