Many-Valued logics

($\alpha$-)SAT solver

Automated theorem ($\alpha$-)prover

Many-Valued logics

In addition to sat and prove, SoleReasoners also provides methods to deal with their many-valued couterparts, respectively alphasat and alphaprove. The former aims at solving the $\alpha$-satisfiability problem, i.e., if there exists a model such that a given many-valued formula assumes value of at least $\alpha$, while the latter aims at solving the $\alpha$-validity problem, i.e., if there is not a model for which a given many-valued formula does not assume value of at least $\alpha$.

For more information about the data structures involved, please refer to the developer documentation.

($\alpha$-)SAT solver

SoleReasoners.sat — Method

sat(
    z::Formula,
    h::A;
    rng = Random.GLOBAL_RNG,
    verbose::Bool=false,
    timeout::Union{Nothing,Int}=nothing,
    kwargs...
) where {
    T<:Truth,
    D<:AbstractVector{T},
    A<:FiniteAlgebra{T,D}
}

Given a formula, return true if an interpretation that satisfies the formula exists, false otherwise.

SoleReasoners.alphasat — Method

alphasat(
    α::T1,
    z::Formula,
    a::A;
    rng = Random.GLOBAL_RNG,
    verbose::Bool=false,
    timeout::Union{Nothing,Int}=nothing,
    kwargs...
) where {
    T<:Truth,
    D<:AbstractVector{T},
    A<:FiniteAlgebra{T,D},
    T1<:Truth
}

Given a formula, return true if it is α-satisfiable, i.e., there is an interpretation such that the formula assumes value of at least α, nothing in case of timeout or out-of-memory error, false otherwise.

Automated theorem ($\alpha$-)prover

SoleReasoners.prove — Method

prove(
    z::Formula,
    h::A;
    rng = Random.GLOBAL_RNG,
    verbose::Bool=false,
    timeout::Union{Nothing,Int}=nothing,
    kwargs...
) where {
    T<:Truth,
    D<:AbstractVector{T},
    A<:FiniteAlgebra{T,D}
}

Given a formula, return true if it is valid, i.e., there is not an interpretation that does not satisfy the formula, false otherwise.

SoleReasoners.alphaprove — Method

alphaprove()
    α::T1,
    z::Formula,
    a::A;
    rng = Random.GLOBAL_RNG,
    verbose::Bool=false,
    timeout::Union{Nothing,Int}=nothing,
    kwargs...
) where {
    T<:Truth,
    D<:AbstractVector{T},
    A<:FiniteAlgebra{T,D},
    T1<:Truth
}

Given a formula, return true if it is α-valid, i.e., there is not an interpretation such that the formula does not assume value of at least α, nothing in case of timeout or out-of-memory error, false otherwise.