Getting started

Tableau

MetricHeapNode

MetricHeapOrdering

MetricHeap

SAT

Utils

Getting started

In this introductory section you will learn about the main building blocks of SoleReasoners. Their definition, usage examples and how to customize them to your own needings.

Tableau

SoleReasoners.Tableau — Type

struct Tableau
    φ::Formula
    father::Base.RefValue{Set{Tableau}}
    children::Base.RefValue{Set{Tableau}}
    literals::Base.RefValue{Set{Formula}}
end

A recursive structure resembling a tree which contains a formula φ (which structure is defined in the SoleLogics package) and references to the father, the set of children and the set of literals (atoms and negation of atoms) found in the same branch.

A tableau is a dynamic structure used to solve the satisfiability problem following the analytic tableau approach. At a specific instant in time, the root represents a possible expansion node of the tableau, while each path from a root to one of its leaves represents a branch, with literals in the leaves containing the literals (atoms and negation of atoms) encountered it the expanded nodes for that branch.

This allows to delegate to the Julia garbage collector the deallocation of expanded nodes, while simplifying root (expansion node) search and branch management.

Note that, while at initialization only there is only one root, the structure later on can be divided in many tree-like structures, with each root representing an expansion node and each set of leaves representing the leaves (or branches) associated with that expansion node.

SoleReasoners.φ — Method

Getter for the formula of a tableau.

SoleReasoners.fatherset — Method

Getter for the set containing the father of a tableau.

SoleReasoners.father — Method

Getter for the father of a tableau.

SoleReasoners.childrenset — Method

Getter for the set containing the children of a tableau.

SoleReasoners.literals — Method

Getter for the set containing the literals of a tableau.

SoleReasoners.leaves — Method

Getter for the leaves of a tableau.

SoleReasoners.leaves — Method

Getter for the leaves of a tableau.

SoleReasoners.pushfather! — Method

Push new father to a tableau.

SoleReasoners.popfather! — Method

Pop father of a tableau (the tableau becomes a root).

SoleReasoners.pushchildren! — Method

Push new children to a tableau.

SoleReasoners.pushliterals! — Method

Push new literals to a tableau.

SoleReasoners.findroot — Method

Find root starting from the leaf (i.e., the expansion node relative to that leaf).

SoleReasoners.isleaf — Method

Return true if the tableau is still a leaf, false otherwise.

MetricHeapNode

SoleReasoners.MetricHeapNode — Type

struct MetricHeapNode
    metricvalue::Int
    tableau::Tableau
end

The atomic element of a MetricHeap, it contains a tableau branch and a value for the metric associated with the MetricHeap it is contained in.

SoleReasoners.metricvalue — Method

Getter for the metric value of a heap node.

SoleReasoners.tableau — Method

Getter for the tableau branch of a heap node.

MetricHeapOrdering

SoleReasoners.MetricHeapOrdering — Type

Definition of a new ordering for the heaps treating them as min heaps ordered on the metric value.

Base.Order.lt — Method

Definition of the lt function for the new ordering.

MetricHeap

SoleReasoners.MetricHeap — Type

struct MetricHeap
    heap::BinaryHeap{MetricHeapNode}
    metric::Function
end

A MetricHeap is basically a heap parametrized over a metric, i.e., a function which extracts some information about a tableau branch, therefore containing in each node a tableau branch and the relative value for the metric, and which is ordered as a min heap over this metric value.

SoleReasoners.heap — Method

Getter for the binary heap of a MetricHeap.

SoleReasoners.metric — Method

Getter for the metric function of a MetricHeap.

Base.push! — Method

Push new metricheapnode to a MetricHeap.

Base.push! — Method

Push new tableau to a MetricHeap.

Base.pop! — Method

Pop head of a MetricHeap and return the tableau associated with it.

Base.isempty — Method

Returns true if the MetricHeap is empty, false otherwise.

SAT

SoleReasoners.naivechooseleaf — Method

Choose a leaf using the provided metric heaps. At this moment, it simply returns the leaf which compares the most as head of the heaps.

To prevent starvation, use roundrobin instead.

SoleReasoners.naivechooseleaf — Method

Choose a leaf using the provided metric heaps. At this moment, it simply returns the leaf which compares the most as head of the heaps.

To prevent starvation, use roundrobin instead.

SoleReasoners.roundrobin — Method

Choose a leaf using the provided metric heaps, alternating between them at each cycle.

Base.push! — Method

Push leaf to each metric heap.

Base.push! — Method

Push leaf to each metric heap.

SoleReasoners.sat — Method

Given a formula, return true if an interpretation that satisfies the formula exists, false otherwise.

SoleReasoners.sat — Method

Given a formula, return true if an interpretation that satisfies the formula exists, false otherwise.

SoleReasoners.sat — Method

Given a formula, return true if an interpretation that satisfies the formula exists, false otherwise.

SoleReasoners.sat — Method

Given a formula, return true if an interpretation that satisfies the formula exists, false otherwise.

Utils

SoleReasoners.dimacstosole — Method

Simple parsing from DIMACS CNF format to a SoleLogics Formula.

dimacscnf is the path of the file containing the formula in DIMACS CNF format.