Getting started

Tableau

MetricHeapNode

MetricHeapOrdering

MetricHeap

SAT

Getting started

In this introductory section you will learn about the main building blocks of SoleReasoners. Their definition, usage examples and how to customize them to your own needings.

Tableau

SoleReasoners.Tableau — Type

struct Tableau
    φ::Formula
    father::Base.RefValue{Set{Tableau}}
    children::Base.RefValue{Set{Tableau}}
    literals::Base.RefValue{Set{Formula}}
end

A recursive structure resembling a tree which contains a formula φ (which structure is defined in the SoleLogics package) and references to the father, the set of children and the set of literals (atoms and negation of atoms) found in the same branch.

A tableau is a dynamic structure used to solve the satisfiability problem following the analytic tableau approach. At a specific instant in time, the root represents a possible expansion node of the tableau, while each path from a root to one of its leaves represents a branch, with literals in the leaves containing the literals (atoms and negation of atoms) encountered it the expanded nodes for that branch.

This allows to delegate to the Julia garbage collector the deallocation of expanded nodes, while simplifying root (expansion node) search and branch management.

Note that, while at initialization only there is only one root, the structure later on can be divided in many tree-like structures, with each root representing an expansion node and each set of leaves representing the leaves (or branches) associated with that expansion node.

SoleReasoners.φ — Method

φ(tableau::Tableau)::Formula = tableau.φ

Getter for the formula of a tableau.

SoleReasoners.fatherset — Method

fatherset(tableau::Tableau)::Set{Tableau} = tableau.father[]

Getter for the set containing the father of a tableau.

SoleReasoners.father — Method

father(tableau::Tableau)::Union{Tableau, Nothing}

Getter for the father of a tableau.

SoleReasoners.childrenset — Method

childrenset(tableau::Tableau)::Set{Tableau}

Getter for the set containing the children of a tableau.

SoleReasoners.literals — Method

literals(tableau::Tableau)::Set{Formula}

Getter for the set containing the literals of a tableau.

SoleReasoners.leaves — Method

leaves(leavesset::Set{Tableau}, tableau::Tableau)::Set{Tableau}

Getter for the leaves of a tableau.

SoleReasoners.leaves — Method

leaves(tableau::Tableau)::Set{Tableau}

Getter for the leaves of a tableau.

SoleReasoners.pushfather! — Method

pushfather!(tableau::Tableau, newfather::Tableau)::Set{Tableau}

Push new father to a tableau.

SoleReasoners.popfather! — Method

popfather!(tableau::Tableau)::Tableau

Pop father of a tableau (the tableau becomes a root).

SoleReasoners.pushchildren! — Method

pushchildren!(tableau::Tableau, children::Tableau...)::Set{Tableau}

Push new children to a tableau.

SoleReasoners.pushliterals! — Method

pushliterals!(tableau::Tableau, newliterals::Formula...)::Set{Formula}

Push new literals to a tableau.

SoleReasoners.findroot — Method

findroot(tableau::Tableau)::Tableau

Find root starting from the leaf (i.e., the expansion node relative to that leaf).

SoleReasoners.isleaf — Method

isleaf(tableau::Tableau)::Bool

Return true if the tableau is still a leaf, false otherwise.

MetricHeapNode

SoleReasoners.MetricHeapNode — Type

struct MetricHeapNode
    metricvalue::Int
    tableau::Tableau
end

The atomic element of a MetricHeap, it contains a tableau branch and a value for the metric associated with the MetricHeap it is contained in.

SoleReasoners.metricvalue — Method

metricvalue(metricheapnode::MetricHeapNode)

Getter for the metric value of a heap node.

SoleReasoners.tableau — Method

tableau(metricheapnode::MetricHeapNode)

Getter for the tableau branch of a heap node.

MetricHeapOrdering

SoleReasoners.MetricHeapOrdering — Type

struct MetricHeapOrdering <: Base.Order.Ordering end

Definition of a new ordering for the heaps treating them as min heaps ordered on the metric value.

Base.Order.lt — Method

lt(o::MetricHeapOrdering, a::MetricHeapNode, b::MetricHeapNode)::Bool

Definition of the lt function for the new ordering.

MetricHeap

SoleReasoners.MetricHeap — Type

struct MetricHeap
    heap::BinaryHeap{MetricHeapNode}
    metric::Function
end

A MetricHeap is basically a heap parametrized over a metric, i.e., a function which extracts some information about a tableau branch, therefore containing in each node a tableau branch and the relative value for the metric, and which is ordered as a min heap over this metric value.

SoleReasoners.heap — Method

heap(metricheap::MetricHeap)::BinaryHeap{MetricHeapNode}

Getter for the binary heap of a MetricHeap.

SoleReasoners.metric — Method

metric(metricheap::MetricHeap)::Function

Getter for the metric function of a MetricHeap.

Base.push! — Method

push!(metricheap::MetricHeap, metricheapnode::MetricHeapNode)::BinaryHeap{MetricHeapNode}

Push new metricheapnode to a MetricHeap.

Base.push! — Method

push!(metricheap::MetricHeap, tableau::Tableau)::BinaryHeap{MetricHeapNode}

Push new tableau to a MetricHeap.

Base.pop! — Method

pop!(metricheap::MetricHeap)::Tableau

Pop head of a MetricHeap and return the tableau associated with it.

Base.isempty — Method

isempty(metricheap::MetricHeap)::Bool

Returns true if the MetricHeap is empty, false otherwise.

SAT

SoleReasoners.naivechooseleaf — Method

naivechooseleaf(metricheaps::Vector{MetricHeap})::Union{Tableau, Nothing}

Choose a leaf using the provided metric heaps. At this moment, it simply returns the leaf which compares the most as head of the heaps.

To prevent starvation, use roundrobin instead.

SoleReasoners.naivechooseleaf — Method

naivechooseleaf(metricheaps::Vector{MetricHeap}, _::Int)

Choose a leaf using the provided metric heaps. At this moment, it simply returns the leaf which compares the most as head of the heaps.

To prevent starvation, use roundrobin instead.

SoleReasoners.roundrobin — Method

roundrobin(metricheaps::Vector{MetricHeap}, cycle::Int)::Union{Tableau, Nothing}

Choose a leaf using the provided metric heaps, alternating between them at each cycle.

Base.push! — Method

push!(metricheaps::Vector{MetricHeap}, tableau::Tableau)::Nothing

Push leaf to each metric heap.

SoleReasoners.sat — Method

sat(metricheaps::Vector{MetricHeap}, chooseleaf::Function)::Bool

Given a formula, return true if an interpretation that satisfies the formula exists, false otherwise.

SoleReasoners.sat — Method

sat(φ::Formula, chooseleaf::Function, metrics::Function...)::Bool

Given a formula, return true if an interpretation that satisfies the formula exists, false otherwise.

SoleReasoners.sat — Method

sat(φ::Formula, chooseleaf::Function; rng = Random.GLOBAL_RNG)::Bool

Given a formula, return true if an interpretation that satisfies the formula exists, false otherwise.

SoleReasoners.sat — Method

sat(φ::Formula; rng = Random.GLOBAL_RNG)::Bool

Given a formula, return true if an interpretation that satisfies the formula exists, false otherwise.