Getting started

In this introductory section you will learn about the main building blocks of SoleReasoners. Their definition, usage examples and how to customize them to your own needings.

Tableau

SoleReasoners.Tableau — Type

struct Tableau
    φ::Formula
    father::Base.RefValue{Set{Tableau}}
    children::Base.RefValue{Set{Tableau}}
    literals::Base.RefValue{Set{Formula}}
end

A recursive structure resembling a tree which contains a formula φ (which structure is defined in the SoleLogics package) and references to the father, the set of children and the set of literals (atoms and negation of atoms) found in the same branch.

A tableau is a dynamic structure used to solve the satisfiability problem following the analytic tableau approach. At a specific instant in time, the root represents a possible expansion node of the tableau, while each path from a root to one of its leaves represents a branch, with literals in the leaves containing the literals (atoms and negation of atoms) encountered it the expanded nodes for that branch.

This allows to delegate to the Julia garbage collector the deallocation of expanded nodes, while simplifying root (expansion node) search and branch management.

Note that, while at initialization only there is only one root, the structure later on can be divided in many tree-like structures, with each root representing an expansion node and each set of leaves representing the leaves (or branches) associated with that expansion node.

Missing docstring.

Missing docstring for φ(tableau::Tableau)::Formula = tableau.φ. Check Documenter's build log for details.

Missing docstring.

Missing docstring for fatherset(tableau::Tableau)::Set{Tableau}. Check Documenter's build log for details.

Missing docstring.

Missing docstring for father(tableau::Tableau)::Union{Tableau, Nothing}. Check Documenter's build log for details.

Missing docstring.

Missing docstring for childrenset(tableau::Tableau)::Set{Tableau}. Check Documenter's build log for details.

Missing docstring.

Missing docstring for literals(tableau::Tableau)::Set{Formula}. Check Documenter's build log for details.

Missing docstring.

Missing docstring for leaves(leavesset::Set{Tableau}, tableau::Tableau)::Set{Tableau}. Check Documenter's build log for details.

Missing docstring.

Missing docstring for leaves(tableau::Tableau)::Set{Tableau}. Check Documenter's build log for details.

Missing docstring.

Missing docstring for pushfather!(tableau::Tableau, newfather::Tableau)::Set{Tableau}. Check Documenter's build log for details.

Missing docstring.

Missing docstring for popfather!(tableau::Tableau)::Tableau. Check Documenter's build log for details.

Missing docstring.

Missing docstring for pushchildren!(tableau::Tableau, children::Tableau...)::Set{Tableau}. Check Documenter's build log for details.

Missing docstring.

Missing docstring for pushliterals!(tableau::Tableau, newliterals::Formula...)::Set{Formula}. Check Documenter's build log for details.

Missing docstring.

Missing docstring for findroot(tableau::Tableau)::Tableau. Check Documenter's build log for details.

Missing docstring.

Missing docstring for isleaf(tableau::Tableau)::Bool. Check Documenter's build log for details.

MetricHeapNode

SoleReasoners.MetricHeapNode — Type

struct MetricHeapNode
    metricvalue::Int
    tableau::Tableau
end

The atomic element of a MetricHeap, it contains a tableau branch and a value for the metric associated with the MetricHeap it is contained in.

SoleReasoners.metricvalue — Method

Getter for the metric value of a heap node.

SoleReasoners.tableau — Method

Getter for the tableau branch of a heap node.

MetricHeapOrdering

SoleReasoners.MetricHeapOrdering — Type

Definition of a new ordering for the heaps treating them as min heaps ordered on the metric value.

Missing docstring.

Missing docstring for lt(o::MetricHeapOrdering, a::MetricHeapNode, b::MetricHeapNode)::Bool. Check Documenter's build log for details.

MetricHeap

SoleReasoners.MetricHeap — Type

struct MetricHeap
    heap::BinaryHeap{MetricHeapNode}
    metric::Function
end

A MetricHeap is basically a heap parametrized over a metric, i.e., a function which extracts some information about a tableau branch, therefore containing in each node a tableau branch and the relative value for the metric, and which is ordered as a min heap over this metric value.

Missing docstring.

Missing docstring for heap(metricheap::MetricHeap)::BinaryHeap{MetricHeapNode}. Check Documenter's build log for details.

Missing docstring.

Missing docstring for metric(metricheap::MetricHeap)::Function. Check Documenter's build log for details.

Missing docstring.

Missing docstring for push!(metricheap::MetricHeap, metricheapnode::MetricHeapNode)::BinaryHeap{MetricHeapNode}. Check Documenter's build log for details.

Missing docstring.

Missing docstring for push!(metricheap::MetricHeap, tableau::Tableau)::BinaryHeap{MetricHeapNode}. Check Documenter's build log for details.

Missing docstring.

Missing docstring for pop!(metricheap::MetricHeap)::Tableau. Check Documenter's build log for details.

Missing docstring.

Missing docstring for isempty(metricheap::MetricHeap)::Bool. Check Documenter's build log for details.

SAT

Missing docstring.

Missing docstring for naivechooseleaf(metricheaps::Vector{MetricHeap})::Union{Tableau, Nothing}. Check Documenter's build log for details.

Missing docstring.

Missing docstring for naivechooseleaf(metricheaps::Vector{MetricHeap}, cycle::Int). Check Documenter's build log for details.

Missing docstring.

Missing docstring for roundrobin(metricheaps::Vector{MetricHeap}, cycle::Int)::Union{Tableau, Nothing}. Check Documenter's build log for details.

Missing docstring.

Missing docstring for push!(metricheaps::Set{MetricHeap}, tableau::Tableau)::Nothing. Check Documenter's build log for details.

Missing docstring.

Missing docstring for push!(metricheaps::Vector{MetricHeap}, tableau::Tableau)::Nothing. Check Documenter's build log for details.

Missing docstring.

Missing docstring for sat(metricheaps::Vector{MetricHeap}, chooseleaf::Function)::Bool. Check Documenter's build log for details.

Missing docstring.

Missing docstring for sat(φ::Formula, chooseleaf::Function, metrics::Function...)::Bool. Check Documenter's build log for details.

Missing docstring.

Missing docstring for sat(φ::Formula, chooseleaf::Function; rng = Random.GLOBAL_RNG)::Bool. Check Documenter's build log for details.

Missing docstring.

Missing docstring for sat(φ::Formula; rng = Random.GLOBAL_RNG)::Bool. Check Documenter's build log for details.

Utils

Missing docstring.

Missing docstring for dimacstosole(dimacscnf::String)::Formula. Check Documenter's build log for details.