Getting started

SAT solver

Automated theorem prover

Getting started

SoleReasoners mainly provides two tools for reasoning: sat, which aims at solving the boolean satisfiability problem, and prove, which serves as an automated theorem prover for propositional boolean formulae. Both algorithms are based on the method of analytic tableaux.

SoleReasoners also provides a suite for managing the tableau expansion policy, allowing for different configurations based on the specific application problem. This system is based on min-heaps concurring to provide candidates for the extraction based on user specified policies, and an agreement function to choose amongst such proposals.

For more information about the data structures involved, please refer to the developer documentation.

SAT solver

SoleReasoners.sat — Method

sat(formula::Formula, choosenode::Function, metrics::Function...)

Given a formula, return true if an interpretation that satisfies the formula exists, false otherwise.

choosenode should be a function taking a vector of metricheaps as an argument (and eventually a counter) and giving a tableau (or nothing) as output that is used to extract a node representing a branch to be expanded. If nothing, all branches are closed.

metrics should be functions taking a tableau as an argument and giving an integer as output that are used to model the order in which tableau branches are xpanded. For example, one could declare the following metric functions:

mf1(t::Tableau) = noperators(t.formula)
mf2(t::Tableau) = height(t.formula)

The first metric will generate a metricheap proposing to expand first branches comprising the node containing the formula with the less number of operators, the second metric will generate a metricheap proposing to expand first branches comprising the node containing the formula of less height.

choosenode will then be used to choose which policy to follow (e.g., choosing the node voted by most heaps, or alternating between each heap at each cycle).

SoleReasoners.sat — Method

sat(
    formula::Formula,
    metric::F;
    rng = Random.GLOBAL_RNG
) where {
    F<:Function
}

Given a formula and an extraction policy metric, return true if an interpretation that satisfies the formula exists, false otherwise.

metric should be a function taking a tableau as an argument and giving an integer as output that is used to model the order in which tableau branches are expanded. For example, one could declare the following metric function:

mf(t::Tableau) = noperators(t.formula)

This way, the tableau will be expanded giving precedence to branches comprising nodes containing formulae with the smallest number operators.

SoleReasoners.sat — Method

sat(formula::Formula; rng = Random.GLOBAL_RNG)

Given a formula, return true if an interpretation that satisfies the formula exists, false otherwise.

Automated theorem prover

SoleReasoners.prove — Method

sat(formula::Formula, choosenode::Function, metrics::Function...)

Given a formula, return true if it is valid, i.e., there is not an interpretation that does not satisfy the formula, false otherwise.

choosenode should be a function taking a vector of metricheaps as an argument (and eventually a counter) and giving a tableau (or nothing) as output that is used to extract a node representing a branch to be expanded. If nothing, all branches are closed.

metrics should be functions taking a tableau as an argument and giving an integer as output that are used to model the order in which tableau branches are expanded. For example, one could declare the following metric functions:

mf1(t::Tableau) = noperators(t.formula)
mf2(t::Tableau) = height(t.formula)

The first metric will generate a metricheap proposing to expand first branches comprising the node containing the formula with the less number of operators, the second metric will generate a metricheap proposing to expand first branches omprising the node containing the formula of less height.

choosenode will then be used to choose which policy to follow (e.g., choosing the node voted by most heaps, or alternating between each heap at each cycle).

SoleReasoners.prove — Method

prove(formula::Formula, metric::F; kwargs...) where {F<:Function}

Given a formula, return true if it is valid, i.e., there is not an interpretation that does not satisfy the formula, false otherwise.

metric should be a function taking a tableau as an argument and giving an integer as output that is used to model the order in which tableau branches are expanded. For example, one could declare the following metric function:

mf(t::Tableau) = noperators(t.formula)

This way, the tableau will be expanded giving precedence to branches comprising nodes containing formulae with the smallest number operators.

SoleReasoners.prove — Method

prove(formula::Formula; rng = Random.GLOBAL_RNG)

Given a formula, return true if it is valid, i.e., there is not an interpretation that does not satisfy the formula, false otherwise.